pyproject.toml
[project]
name = "jambashrimp"
version = "0.1.0"
description = "JambaShrimp (jash): playful CLI to chat with PDFs or crawled docs using Jamba Reasoning 3B"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "httpx>=0.27.2",
    "huggingface-hub>=0.36.0",
    "llama-cpp-python>=0.2.92",
    "pypdf>=6.4.0",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "prompt-toolkit>=3.0.0",
    "lxml>=5.3.0",
    "numpy>=1.26.0",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.1",
]

[project.scripts]
jash = "jamba_cli.cli:main"
jamba-chat = "jamba_cli.cli:main"
jash-setup = "jamba_cli.setup_assets:main"

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true
